requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.2.0
mysqlclient==2.2.4
django-allauth==0.61.1
django-rest-auth==0.9.5
python-dotenv==1.0.1
redis==5.0.1
django-redis==5.4.0
django-debug-toolbar==4.3.0
black==24.2.0
flake8==7.0.0
celery==5.3.6
pytest==8.0.2
pytest-django==4.8.0
gunicorn==21.2.0
whitenoise==6.6.0
